Ensure to convert pod status to juju status consistent across operator and sidecar applications;
