+++ caas/kubernetes/provider/resources/clusterrolebinding_test.go
@@ -11,7 +11,6 @@ import (
 	gc "gopkg.in/check.v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/juju/juju/caas/kubernetes/provider/resources"
@@ -119,7 +118,7 @@ func (s *clusterRoleBindingSuite) TestDeleteWithoutPreconditions(c *gc.C) {
 // https://bugs.launchpad.net/juju/+bug/1929909
 func (s *clusterRoleBindingSuite) TestEnsureClusterRoleBindingRegressionOnLabelChange(c *gc.C) {
 	clusterRoleBinding := &rbacv1.ClusterRoleBinding{
-		ObjectMeta: meta.ObjectMeta{
+		ObjectMeta: metav1.ObjectMeta{
 			Name: "test",
 			Labels: map[string]string{
 				"foo": "bar",
@@ -138,7 +137,7 @@ func (s *clusterRoleBindingSuite) TestEnsureClusterRoleBindingRegressionOnLabelC
 	rroleBinding, err := s.client.RbacV1().ClusterRoleBindings().Get(
 		context.TODO(),
 		"test",
-		meta.GetOptions{},
+		metav1.GetOptions{},
 	)
 
 	c.Assert(err, jc.ErrorIsNil)
@@ -158,7 +157,7 @@ func (s *clusterRoleBindingSuite) TestEnsureClusterRoleBindingRegressionOnLabelC
 	rroleBinding, err = s.client.RbacV1().ClusterRoleBindings().Get(
 		context.TODO(),
 		"test",
-		meta.GetOptions{},
+		metav1.GetOptions{},
 	)
 
 	c.Assert(err, jc.ErrorIsNil)
@@ -169,7 +168,7 @@ func (s *clusterRoleBindingSuite) TestEnsureRecreatesOnRoleRefChange(c *gc.C) {
 	clusterRoleBinding := resources.NewClusterRoleBinding(
 		"test",
 		&rbacv1.ClusterRoleBinding{
-			ObjectMeta: meta.ObjectMeta{
+			ObjectMeta: metav1.ObjectMeta{
 				Name: "test",
 				Labels: map[string]string{
 					"foo": "bar",
@@ -202,7 +201,7 @@ func (s *clusterRoleBindingSuite) TestEnsureRecreatesOnRoleRefChange(c *gc.C) {
 	rval, err := s.client.RbacV1().ClusterRoleBindings().Get(
 		context.TODO(),
 		"test",
-		meta.GetOptions{},
+		metav1.GetOptions{},
 	)
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(rval.ObjectMeta.ResourceVersion, gc.Equals, "1")
@@ -210,7 +209,7 @@ func (s *clusterRoleBindingSuite) TestEnsureRecreatesOnRoleRefChange(c *gc.C) {
 	clusterRoleBinding1 := resources.NewClusterRoleBinding(
 		"test",
 		&rbacv1.ClusterRoleBinding{
-			ObjectMeta: meta.ObjectMeta{
+			ObjectMeta: metav1.ObjectMeta{
 				Name: "test",
 				Labels: map[string]string{
 					"foo": "bar",
