+++ caas/kubernetes/provider/resources/events.go
@@ -15,9 +15,21 @@ import (
 
 const maxEventsToPage = 100
 
+func listEventsForObject(
+	ctx context.Context, client kubernetes.Interface, namespace, name, kind string,
+) ([]corev1.Event, error) {
+	return ListEventsForObject(ctx, name, kind, client.CoreV1().Events(namespace))
+}
+
+// EventsGetter defines methods for fetching k8s events.
+type EventsGetter interface {
+	List(context.Context, metav1.ListOptions) (*corev1.EventList, error)
+}
+
 // ListEventsForObject returns all the events for the specified object.
-func ListEventsForObject(ctx context.Context, client kubernetes.Interface,
-	namespace string, name string, kind string) ([]corev1.Event, error) {
+func ListEventsForObject(
+	ctx context.Context, name string, kind string, getEvents EventsGetter,
+) ([]corev1.Event, error) {
 	selector := fields.AndSelectors(
 		fields.OneTermEqualSelector("involvedObject.name", name),
 		fields.OneTermEqualSelector("involvedObject.kind", kind),
@@ -25,10 +37,9 @@ func ListEventsForObject(ctx context.Context, client kubernetes.Interface,
 	opts := metav1.ListOptions{
 		FieldSelector: selector,
 	}
-	api := client.CoreV1().Events(namespace)
 	var items []corev1.Event
 	for len(items) < maxEventsToPage {
-		res, err := api.List(ctx, opts)
+		res, err := getEvents.List(ctx, opts)
 		if err != nil {
 			return nil, errors.Trace(err)
 		}
