+++ caas/kubernetes/provider/resources/events_test.go
@@ -40,7 +40,7 @@ func (s *eventsSuite) TestList(c *gc.C) {
 		c.Assert(err, jc.ErrorIsNil)
 		res = append(res, ev)
 	}
-	events, err := resources.ListEventsForObject(context.TODO(), s.client, "test", "test", "Pod")
+	events, err := resources.ListEventsForObject(context.TODO(), "test", "Pod", s.client.CoreV1().Events("test"))
 	c.Assert(err, jc.ErrorIsNil)
 
 	toInt := func(s string) int {
